# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Public API documentation served through QRUPI API Gateway.",
        "title": "QRUPI V2 API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/districts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get all districts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by regency code",
                        "name": "regency_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search district name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "code",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/districts/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get district by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "District code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/provinces": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get all provinces",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search province name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "code",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/provinces/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get province by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/regencies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get all regencies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by province code",
                        "name": "province_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search regency name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "code",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/regencies/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get regency by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Regency code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/villages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get all villages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by district code",
                        "name": "district_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search village name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "code",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/villages/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Region API"
                ],
                "summary": "Get village by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Village code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}